fix route

+++ src/app/api/auth/[...nextauth]/actionPicture.ts
@@ -1,5 +1,8 @@
 "use server"
 
+import { getServerSession } from "next-auth";
+import { authConfig } from "./config";
+
 export async function getPicture() {
     const response = await fetch(`${process.env.FETCH_URL}/api/picture`)
     const result: string[] = await response.json();
@@ -7,6 +10,9 @@ export async function getPicture() {
 }
 
 export async function addPicture(data: FormData) {
+    const session = await getServerSession(authConfig);
+    if (!session) return { message: "Access closed" };
+
     const response = await fetch(`${process.env.FETCH_URL}/api/picture`, {
         method: "POST",
         body: data
@@ -16,7 +22,12 @@ export async function addPicture(data: FormData) {
 }
 
 export async function delPicture(img: string) {
-    const response = await fetch(`${process.env.FETCH_URL}/picture?d=${img}`, {
+    const session = await getServerSession(authConfig);
+    if (!session) return { message: "Access closed" };
+    console.log("actionDelete", img);
+    
+
+    const response = await fetch(`${process.env.FETCH_URL}/api/picture?d=${img}`, {
         method: "DELETE"
     })
     const result = await response.json();
