+++ src/app/api/auth/[...nextauth]/actionStock.ts
@@ -1,8 +1,10 @@
 "use server"
 
 import type { Stock } from "@prisma/client";
+import { getServerSession } from "next-auth";
+import { authConfig } from "./config";
 
-export async function getAllStocks() {   
+export async function getAllStocks() {
     const response = await fetch(`${process.env.FETCH_URL}/api/stock`
     );
     const result: Stock[] = await response.json();
@@ -10,6 +12,9 @@ export async function getAllStocks() {
 }
 
 export async function createStock(data: FormData) {
+    const session = await getServerSession(authConfig);
+    if (!session) return { message: "Access closed" };
+
     const response = await fetch(`${process.env.FETCH_URL}/api/stock`, {
         method: "POST",
         body: data
@@ -18,7 +23,10 @@ export async function createStock(data: FormData) {
     return result;
 }
 
-export async function updateStock(data: FormData) {   
+export async function updateStock(data: FormData) {
+    const session = await getServerSession(authConfig);
+    if (!session) return { message: "Access closed" };
+
     const response = await fetch(`${process.env.FETCH_URL}/api/stock`, {
         method: "PUT",
         body: data
@@ -28,6 +36,9 @@ export async function updateStock(data: FormData) {
 }
 
 export async function removeStock(data: Stock) {
+    const session = await getServerSession(authConfig);
+    if (!session) return { message: "Access closed" };
+
     const response = await fetch(`${process.env.FETCH_URL}/api/stock`, {
         method: "DELETE",
         body: JSON.stringify(data)
